make requant lib a separate file
